[IMP] Moved response_success out of util.py into blueprints namespace
