+++ source/app/blueprints/rest/alerts_routes.py
@@ -59,7 +59,7 @@ from app.schema.marshables import IocSchema
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
 from app.util import add_obj_history_entry
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 alerts_rest_blueprint = Blueprint('alerts_rest', __name__)
