+++ source/app/blueprints/rest/api_routes.py
@@ -20,7 +20,7 @@ from flask import Blueprint
 
 from app import app
 from app.blueprints.access_controls import ac_api_requires
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 rest_api_blueprint = Blueprint('rest_api', __name__)
