+++ source/app/blueprints/rest/case/case_assets_routes.py
@@ -58,7 +58,7 @@ from app.schema.marshables import CommentSchema
 from app.blueprints.access_controls import ac_requires_case_identifier
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 from app.blueprints.access_controls import ac_api_return_access_denied
 
 case_assets_rest_blueprint = Blueprint('case_assets_rest', __name__)
