+++ source/app/blueprints/rest/case/case_graphs_routes.py
@@ -26,7 +26,7 @@ from app.datamgmt.case.case_events_db import get_case_events_ioc_graph
 from app.models.authorization import CaseAccessLevel
 from app.blueprints.access_controls import ac_requires_case_identifier
 from app.blueprints.access_controls import ac_api_requires
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 case_graph_rest_blueprint = Blueprint('case_graph_rest', __name__)
