+++ source/app/blueprints/rest/case/case_ioc_routes.py
@@ -55,7 +55,7 @@ from app.blueprints.access_controls import ac_requires_case_identifier
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.access_controls import ac_api_return_access_denied
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 case_ioc_rest_blueprint = Blueprint('case_ioc_rest', __name__)
