+++ source/app/blueprints/rest/case/case_routes.py
@@ -72,7 +72,7 @@ from app.blueprints.access_controls import ac_api_requires
 from app.util import add_obj_history_entry
 from app.blueprints.access_controls import ac_api_return_access_denied
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 case_rest_blueprint = Blueprint('case_rest', __name__)
