+++ source/app/blueprints/rest/case_comments.py
@@ -22,7 +22,7 @@ from flask import request
 
 from app.schema.marshables import CommentSchema
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 from app.business.case_comments import case_comments_update
 from app.business.errors import BusinessProcessingError
