+++ source/app/blueprints/rest/dashboard_routes.py
@@ -46,7 +46,7 @@ from app.blueprints.access_controls import ac_requires_case_identifier
 from app.blueprints.access_controls import ac_api_requires
 from app.util import not_authenticated_redirection_url
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 from app.util import is_authentication_oidc
 
 from oic.oauth2.exception import GrantError
