+++ source/app/blueprints/rest/datastore_routes.py
@@ -50,7 +50,7 @@ from app.blueprints.access_controls import ac_requires_case_identifier
 from app.blueprints.access_controls import ac_api_requires
 from app.util import add_obj_history_entry
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 datastore_rest_blueprint = Blueprint('datastore_rest', __name__)
