+++ source/app/blueprints/rest/filters_routes.py
@@ -26,7 +26,7 @@ from app.datamgmt.filters.filters_db import list_filters_by_type
 from app.iris_engine.utils.tracker import track_activity
 from app.schema.marshables import SavedFilterSchema
 from app.blueprints.access_controls import ac_api_requires
-from app.util import response_success
+from app.blueprints.responses import response_success
 from app.blueprints.responses import response_error
 
 saved_filters_rest_blueprint = Blueprint('saved_filters_rest', __name__)
