+++ source/app/blueprints/rest/manage/manage_access_control_routes.py
@@ -25,7 +25,7 @@ from app.iris_engine.access_control.utils import ac_trace_effective_user_permiss
 from app.iris_engine.access_control.utils import ac_trace_user_effective_cases_access_2
 from app.models.authorization import Permissions
 from app.blueprints.access_controls import ac_api_requires
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 manage_ac_rest_blueprint = Blueprint('access_control_rest', __name__)
