+++ source/app/blueprints/rest/manage/manage_alerts_status_routes.py
@@ -27,7 +27,7 @@ from app.schema.marshables import AlertStatusSchema
 from app.schema.marshables import AlertResolutionSchema
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 manage_alerts_status_rest_blueprint = Blueprint('manage_alerts_status_rest', __name__)
