+++ source/app/blueprints/rest/manage/manage_assets_type_routes.py
@@ -33,7 +33,7 @@ from app.models.models import CaseAssets
 from app.schema.marshables import AssetTypeSchema
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 manage_assets_type_rest_blueprint = Blueprint('manage_assets_type_rest', __name__)
