+++ source/app/blueprints/rest/manage/manage_attributes_routes.py
@@ -26,7 +26,7 @@ from app.models.authorization import Permissions
 from app.models.models import CustomAttribute
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 manage_attributes_rest_blueprint = Blueprint('manage_attributes_rest', __name__)
