+++ source/app/blueprints/rest/manage/manage_case_templates_routes.py
@@ -34,7 +34,7 @@ from app.schema.marshables import CaseTemplateSchema
 from app.blueprints.access_controls import ac_requires_case_identifier
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 manage_case_templates_rest_blueprint = Blueprint('manage_case_templates_rest', __name__)
