+++ source/app/blueprints/rest/manage/manage_cases_routes.py
@@ -54,7 +54,7 @@ from app.blueprints.access_controls import ac_requires_case_identifier
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.access_controls import ac_api_return_access_denied
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 from app.business.cases import cases_delete
 from app.business.cases import cases_update
 from app.business.cases import cases_create
