+++ source/app/blueprints/rest/manage/manage_customers_routes.py
@@ -45,7 +45,7 @@ from app.schema.marshables import ContactSchema
 from app.schema.marshables import CustomerSchema
 from app.blueprints.access_controls import ac_api_requires_client_access
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 manage_customers_rest_blueprint = Blueprint('manage_customers_rest', __name__)
