+++ source/app/blueprints/rest/manage/manage_event_categories_routes.py
@@ -23,7 +23,7 @@ from app.models.models import EventCategory
 from app.schema.marshables import EventCategorySchema
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 manage_event_categories_rest_blueprint = Blueprint('manage_event_categories_rest', __name__)
