+++ source/app/blueprints/rest/manage/manage_groups.py
@@ -46,7 +46,7 @@ from app.schema.marshables import AuthorizationGroupSchema
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.access_controls import ac_api_return_access_denied
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 from app.iris_engine.demo_builder import protect_demo_mode_group
 
 manage_groups_rest_blueprint = Blueprint('manage_groups_rest', __name__)
