+++ source/app/blueprints/rest/manage/manage_modules_routes.py
@@ -40,7 +40,7 @@ from app.iris_engine.utils.tracker import track_activity
 from app.models.authorization import Permissions
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 from app.schema.marshables import IrisModuleSchema
 
 manage_modules_rest_blueprint = Blueprint('manage_module_rest', __name__)
