+++ source/app/blueprints/rest/manage/manage_severities_routes.py
@@ -25,7 +25,7 @@ from app.datamgmt.manage.manage_common import search_severity_by_name
 from app.schema.marshables import SeveritySchema
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 manage_severities_rest_blueprint = Blueprint('manage_severities_rest', __name__)
