+++ source/app/blueprints/rest/manage/manage_task_status_routes.py
@@ -21,7 +21,7 @@ from flask import Blueprint
 from app.models.models import TaskStatus
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 
 manage_task_status_rest_blueprint = Blueprint('manage_task_status_rest', __name__)
