+++ source/app/blueprints/rest/manage/manage_users.py
@@ -49,7 +49,7 @@ from app.schema.marshables import UserFullSchema
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.access_controls import ac_api_return_access_denied
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 from app.iris_engine.demo_builder import protect_demo_mode_user
 
 manage_users_rest_blueprint = Blueprint('manage_users_rest', __name__)
