+++ source/app/blueprints/rest/profile_routes.py
@@ -36,7 +36,7 @@ from app.schema.marshables import UserSchema
 from app.schema.marshables import BasicUserSchema
 from app.blueprints.access_controls import ac_api_requires
 from app.blueprints.responses import response_error
-from app.util import response_success
+from app.blueprints.responses import response_success
 from app.blueprints.rest.endpoints import endpoint_removed
 
 profile_rest_blueprint = Blueprint('profile_rest', __name__)
